+++ kmm/src/commonMain/kotlin/fr/beapp/interviews/bicloo/kmm/ServiceLocator.kt
@@ -2,6 +2,7 @@ package fr.beapp.interviews.bicloo.kmm
 
 import fr.beapp.interviews.bicloo.kmm.core.di.DI
 import fr.beapp.interviews.bicloo.kmm.presentation.contract.ContractPresenter
+import fr.beapp.interviews.bicloo.kmm.presentation.preferences.PreferencesPresenter
 import fr.beapp.interviews.bicloo.kmm.presentation.station.StationPresenter
 import org.kodein.di.provider
 
@@ -12,5 +13,8 @@ object ServiceLocator {
 
 	private val contractProvider by DI.provider<ContractPresenter>()
 	fun contractPresenter() = contractProvider.invoke()
+
+	private val preferencesProvider by DI.provider<PreferencesPresenter>()
+	fun preferencesPresenter() = preferencesProvider.invoke()
 
 }