+++ kmm/CHANGELOG.md
@@ -7,6 +7,7 @@
 - `ContractEntity`, `ContractManager` and `ContractPresenter` classes
 - `StationEntity`,`StationManager` and `StationPresenter` classes
 - `ServiceLocator` object
+- `PreferencesPresenter` and `PreferencesManager` class
 
 ### Chore